requirements.txt
psycopg2-binary
fastapi
fastapi[standard]
pillow
